watch-status: support monitoring a git repository
When workflows finish the tracked file will disappear and the workflow pod
will be deleted. Waiting for the pod at that point will not work.

Instead, we let callers provide a path to a file indicating the current
status. When it disappears or its value no longer matches ours we will
quit.