argo-workflow-status: handle missing pods
Once the action finds a workflow that has completed, it needs to look
for the pod for the workflow to get its logs.

At this point, the pod should have been created. But, it's possible that
the pod has since been destroyed. If the pod has been destroyed, there's
no sense in waiting for it to come along later, any new instances of a
similar pod would be for a future run--not the current run.