Update prettier entrypoint.sh to consider dry run (#8)
* Update entrypoint.sh to consider dry run

When dry-run is set to true and we use the --write flag, we expect to
see what files we'll need to reformat without pushing to remote.
However, this requires us to commit the updated changes caused by
prettier. Previously, the placement of the dry run logic was done before
committing the formatted files, resulting to us not seeing the expected
files format. And only using the '--write' flag will not result to a
failure even if there are files that requires formatting.

By moving the logic after committing the prettier files, it will show us
the comparison between two files, without pushing to remote
(given that we use dry flag to true).

* Update description for dry mode